update dashboard settings UI with slider for guess delay; refactor related styles and logic

+++ packages/shared/src/schemas/channel.ts
@@ -2,6 +2,6 @@ import { z } from "zod";
 
 export const settings = z.object({
 	botActiveWhenOffline: z.boolean(),
-	guessDelayTime: z.number().int().min(0).max(60),
+	guessDelayTime: z.number().min(0).max(60),
 	usagePublic: z.boolean(),
 });
